+++ src/index.js
@@ -9,6 +9,7 @@
 const http 						= require("http");
 const https 					= require("https");
 const queryString 				= require("qs");
+const os 						= require("os");
 const kleur						= require("kleur");
 const { match, deprecate }		= require("moleculer").Utils;
 const pkg						= require("../package.json");
@@ -224,6 +225,61 @@ module.exports = {
 
 				return null;
 			}
+		},
+
+		listAliases: {
+			rest: "list-aliases",
+			params: {
+				grouping: { type: "boolean", optional: true, convert: true },
+				withActionSchema: { type: "boolean", optional: true, convert: true }
+			},
+			handler(ctx) {
+				const grouping = !!ctx.params.grouping;
+				const withActionSchema = !!ctx.params.withActionSchema;
+
+				const actionList = withActionSchema ? this.broker.registry.getActionList({}) : null;
+
+				const res = [];
+
+				this.aliases.forEach(alias => {
+					const obj = {
+						actionName: alias.action,
+						path: alias.path,
+						fullPath: alias.fullPath,
+						methods: alias.method,
+						routePath: alias.route.path
+					};
+
+					if (withActionSchema && alias.action) {
+						const actionSchema = actionList.find(item => item.name == alias.action);
+						if (actionSchema && actionSchema.action) {
+							obj.action = _.omit(actionSchema.action, ["handler"]);
+						}
+					}
+
+					if (grouping) {
+						const r = res.find(item => item.route == alias.route);
+						if (r) r.aliases.push(obj);
+						else {
+							res.push({
+								route: alias.route,
+								aliases: [obj]
+							});
+						}
+					} else {
+						res.push(obj);
+					}
+				});
+
+				if (grouping) {
+					res.forEach(item => {
+						item.path = item.route.path;
+						delete item.route;
+					});
+				}
+
+				return res;
+			}
 		}
 	},
 
@@ -244,8 +300,8 @@ module.exports = {
 			}
 
 			// HTTP server timeout
-			if(this.settings.httpServerTimeout){
-				this.logger.debug("Override Default http(s) server timeout:", this.settings.httpServerTimeout);
+			if (this.settings.httpServerTimeout){
+				this.logger.debug("Override default http(s) server timeout:", this.settings.httpServerTimeout);
 				this.server.setTimeout(this.settings.httpServerTimeout);
 			}
 		},
@@ -330,7 +386,7 @@ module.exports = {
 				res.once("finish", () => resolve(true));
 
 				try {
-					await composeThen(req, res, ...route.middlewares);
+					await composeThen.call(this, req, res, ...route.middlewares);
 					let params = {};
 
 					// CORS headers
@@ -577,7 +633,7 @@ module.exports = {
 			} catch(err) {
 				/* istanbul ignore next */
 				if (!err)
-					return;
+					return; // Cancelling promise chain, no error
 
 				throw err;
 			}
@@ -624,18 +680,16 @@ module.exports = {
 				}
 			}
 
-			// Override responseType by action (Deprecated)
+			// Override responseType from action schema
 			let responseType;
 			/* istanbul ignore next */
 			if (action && action.responseType) {
-				deprecate("The 'responseType' action property has been deprecated. Use 'ctx.meta.$responseType' instead");
 				responseType = action.responseType;
 			}
 
-			// Custom headers (Deprecated)
+			// Custom headers from action schema
 			/* istanbul ignore next */
 			if (action && action.responseHeaders) {
-				deprecate("The 'responseHeaders' action property has been deprecated. Use 'ctx.meta.$responseHeaders' instead");
 				Object.keys(action.responseHeaders).forEach(key => {
 					res.setHeader(key, action.responseHeaders[key]);
 					if (key == "Content-Type" && !responseType)
@@ -740,6 +794,7 @@ module.exports = {
 		/**
 		 * Send 404 response
 		 *
+		 * @param {HttpIncomingMessage} req
 		 * @param {HttpResponse} res
 		 */
 		send404(req, res) {
@@ -752,6 +807,7 @@ module.exports = {
 		/**
 		 * Send an error response
 		 *
+		 * @param {HttpIncomingMessage} req
 		 * @param {HttpResponse} res
 		 * @param {Error} err
 		 */
@@ -797,10 +853,21 @@ module.exports = {
 
 			const code = _.isNumber(err.code) && _.inRange(err.code, 400, 599) ? err.code : 500;
 			res.writeHead(code);
-			const errObj = _.pick(err, ["name", "message", "code", "type", "data"]);
-			res.end(JSON.stringify(errObj, null, 2));
+			const errObj = this.reformatError(err, req, res);
+			res.end(errObj !== undefined ? JSON.stringify(errObj, null, 2) : undefined);
 
 			this.logResponse(req, res);
+		},
+
+		/**
+		 * Reformatting the error object to response
+		 * @param {Error} err
+		 * @param {HttpIncomingMessage} req
+		 * @param {HttpResponse} res
+		 @returns {Object}
+		 */
+		reformatError(err/*, req, res*/) {
+			return _.pick(err, ["name", "message", "code", "type", "data"]);
 		},
 
 		/**
@@ -1192,11 +1259,11 @@ module.exports = {
 
 			// `onBeforeCall` handler
 			if (opts.onBeforeCall)
-				route.onBeforeCall = this.Promise.method(opts.onBeforeCall);
+				route.onBeforeCall = opts.onBeforeCall;
 
 			// `onAfterCall` handler
 			if (opts.onAfterCall)
-				route.onAfterCall = this.Promise.method(opts.onAfterCall);
+				route.onAfterCall = opts.onAfterCall;
 
 			// `onError` handler
 			if (opts.onError)
@@ -1397,8 +1464,8 @@ module.exports = {
 			/* istanbul ignore next */
 			this.routes.forEach(route => route.opts.autoAliases && this.regenerateAutoAliases(route));
 		}, 500)
-	},
 
+	},
 
 	events: {
 		"$services.changed"() {
@@ -1420,7 +1487,8 @@ module.exports = {
 					return reject(err);
 
 				const addr = this.server.address();
-				this.logger.info(`API Gateway listening on ${this.isHTTPS ? "https" : "http"}://${addr.address}:${addr.port}`);
+				const listenAddr = addr.address == "0.0.0.0" && os.platform() == "win32" ? "localhost" : addr.address;
+				this.logger.info(`API Gateway listening on ${this.isHTTPS ? "https" : "http"}://${listenAddr}:${addr.port}`);
 				resolve();
 			});
 		});