+++ test/services/test.service.js
@@ -83,22 +83,22 @@ module.exports = {
 				function sleep(ms) {
 					return new Promise(resolve => setTimeout(resolve, ms));
 				}
-				try{
-					var c=0
-					while(c!=ctx.params.counter){
-						await sleep(ctx.params.sleeptime)
+				try {
+					let c = 0;
+					while(c != ctx.params.counter) {
+						await sleep(ctx.params.sleeptime);
 						c++;
 					}
 					return {
-						status:200,
-						msg:"apitimeout response"
-					}
-				}catch(e){
+						status: 200,
+						msg: "apitimeout response"
+					};
+				} catch(e) {
 					return {
-						status:500,
-						msg:"apitimeout response",
-						error:e
-					}
+						status: 500,
+						msg: "apitimeout response",
+						error: e
+					};
 				}
 			}
 		},